fix: financial report, receivable customer ref, income cleanup
- FinancialReport: removed Expense dependency (prepares for expense removal)
- ReceivableResource: fixed broken ->customer reference
- IncomeResource: deleted (table already dropped)
- Integer prices: all monetary columns BIGINT, model casts updated

+++ database/migrations/2025_01_01_000014_create_orders_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('cashier_id')->nullable()->constrained('users')->onDelete('set null');
             $table->enum('status', ['pending', 'diproses', 'selesai'])->default('pending');
             $table->enum('order_type', ['qr', 'cashier'])->default('qr');
-            $table->decimal('total_amount', 15, 2)->default(0);
+            $table->unsignedBigInteger('total_amount')->default(0);
             $table->enum('payment_method', ['cash', 'qris', 'bayar_nanti'])->nullable();
             $table->string('payment_proof', 500)->nullable();
             $table->string('rejection_note', 255)->nullable();
